+++ akash/provider.go
@@ -19,8 +19,9 @@ const ChainId = "chain_id"
 const Node = "node"
 const Home = "home"
 const Path = "path"
+const ProvidersApi = "providers_api"
 
-// Provider -
+// Provider represents the provider resource.
 func Provider() *schema.Provider {
 	return &schema.Provider{
 		Schema: map[string]*schema.Schema{
@@ -72,12 +73,18 @@ func Provider() *schema.Provider {
 				Optional:    true,
 				DefaultFunc: schema.EnvDefaultFunc("AKASH_PATH", "provider-services"),
 			},
+			ProvidersApi: {
+				Type:        schema.TypeString,
+				Optional:    true,
+				DefaultFunc: schema.EnvDefaultFunc("PROVIDERS_API", "http://providers-api.quasarch.cloud"),
+			},
 		},
 		ResourcesMap: map[string]*schema.Resource{
 			"akash_deployment": resourceDeployment(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"akash_deployments": dataSourceDeployments(),
+			"akash_providers":   dataSourceProviders(),
 		},
 		ConfigureContextFunc: providerConfigure,
 	}
@@ -96,6 +103,7 @@ func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}
 		Node:           d.Get(Node).(string),
 		Home:           d.Get(Home).(string),
 		Path:           d.Get(Path).(string),
+		ProvidersApi:   d.Get(ProvidersApi).(string),
 	}
 
 	// Warning or errors can be collected in a slice type
@@ -105,7 +113,7 @@ func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}
 		return nil, diags
 	}
 
-	configuration := client.AkashConfiguration{
+	configuration := client.AkashProviderConfiguration{
 		KeyName:        config[KeyName],
 		KeyringBackend: config[KeyringBackend],
 		AccountAddress: config[AccountAddress],
@@ -115,6 +123,7 @@ func providerConfigure(ctx context.Context, d *schema.ResourceData) (interface{}
 		Node:           config[Node],
 		Home:           config[Home],
 		Path:           config[Path],
+		ProvidersApi:   config[ProvidersApi],
 	}
 
 	tflog.Debug(ctx, fmt.Sprintf("Starting provider with %+v", configuration))