+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] - 2023-01-05
+### Added
+- Add `akash_providers` datasource
+### Changed
+- `akash_deployment` output `services` now has a new structure and more information regarding replicas
+### Fixed
+- Bug where the provider was crashing when no filters were provided in `provider_filters`
+### Development
+- Integration with Praetor's API through custom caching service
+- Start using Akash's type definitions
+- Tracing enabled on deployment creation
+
 ## [0.0.6] - 2022-11-16
 ### Added
 - Add support to mainnet-4